+++ tests/int_allowlist.rs
@@ -184,6 +184,7 @@ fn run_skiplist(skip_regex: Vec<String>, skip_skipword: Vec<String>) -> Result<u
         validation_rps_rule: Vec::new(),
         validation_timeout: 10,
         full_validation_response: false,
+        max_validation_response_length: 2048,
     };
 
     let global_args = GlobalArgs {