+++ tests/int_dedup.rs
@@ -189,6 +189,7 @@ rules:
         validation_rps_rule: Vec::new(),
         validation_timeout: 10,
         full_validation_response: false,
+        max_validation_response_length: 2048,
     };
 
     let global_args = GlobalArgs {