+++ tests/int_teams.rs
@@ -192,6 +192,7 @@ async fn test_scan_teams_messages() -> Result<()> {
         validation_rps_rule: Vec::new(),
         validation_timeout: 10,
         full_validation_response: false,
+        max_validation_response_length: 2048,
     };
 
     let global_args = GlobalArgs {